Remove internal::common
