--- serde_arrow/src/internal/common/utils.rs
@@ -1,2 +1,0 @@
-/// A wrapper type to allow implementing foreign traits
-pub struct Mut<'a, T>(pub &'a mut T);
