+++ serde_arrow/src/internal/deserialization/enum_deserializer.rs
@@ -1,8 +1,8 @@
 use serde::de::{DeserializeSeed, Deserializer, EnumAccess, Visitor};
 
 use crate::internal::{
-    common::Mut,
     error::{fail, Error, Result},
+    utils::Mut,
 };
 
 use super::{array_deserializer::ArrayDeserializer, simple_deserializer::SimpleDeserializer};
