+++ serde_arrow/src/internal/serialization/time_builder.rs
@@ -1,12 +1,11 @@
 use chrono::Timelike;
 
 use crate::internal::{
-    common::MutableBitBuffer,
     error::{Error, Result},
     schema::{GenericField, GenericTimeUnit},
 };
 
-use super::utils::{push_validity, push_validity_default, SimpleSerializer};
+use super::utils::{push_validity, push_validity_default, MutableBitBuffer, SimpleSerializer};
 
 #[derive(Debug, Clone)]
 pub struct TimeBuilder<I> {
