+++ serde_arrow/src/internal/utils.rs
@@ -142,3 +142,6 @@ impl<'a, T: Serialize> Serialize for Items<&'a [T]> {
         seq.end()
     }
 }
+
+/// A wrapper type to allow implementing foreign traits
+pub struct Mut<'a, T>(pub &'a mut T);
